refactor styling and convention typography folder
